+++ NEWS.md
@@ -27,7 +27,7 @@ Red Panda C++ Version 1.5
   - enhancement: When open a project, let user choose weither open it in new window or replace the already openned project
   - fix: editor tooltip for #include_next is not correctly calculated
   - fix: ctrl+click on #include_next header name doesn't open the right file
-  - enhancement: parser used for non-project C files won't find header files in C++ include folders.
+  - enhancement: parser used for non-project C files won't search header files in C++ include folders.
   - fix: toggle block comment/delete to word begin/delete to word end are not correctly disabled when editor not open
 
 Red Panda C++ Version 1.4